MGMT-3510 Set static IP during boot in minimal iso (#986)
* MGMT-3510 Don't include the static IP info in the minimal iso ignition

The configuration will persist from the initramfs to the root
filesystem so we don't want to try to run the unit again.

* MGMT-3510 Move static ip script constant to a separate file

Importing `ignition` into `isoeditor` was causing an import cycle
and nothing in `ignition` uses the script.

The idea here is that any package can import `constants` without
worrying about a cycle because it will never import any other package.

* MGMT-3510 Allow static ip configuration in minimal iso initrd

This commit uses the same script and data we would typically
use in the discovery ignition to configure a static ip before
we try to download the rootfs on first boot.

The script then doesn't need to be provided to the ignition because
the coreos boot process copies the network files from /run/NetworkManager
to /etc/NetworkManager on the rootfs.

ref:
https://github.com/coreos/fedora-coreos-config/blob/076ba7c0917db6d2d2d12c82f0251d5eff32651a/overlay.d/05core/usr/lib/dracut/modules.d/35coreos-ignition/coreos-teardown-initramfs.sh#L69-L104

* MGMT-3510 Add some more info to isoeditor errors