MGMT-3844 Limit concurrent minimal iso generations (#1032)
This commit uses a buffered channel as a semephore to ensure
we're only running 10 minimal iso generation processes concurrently
by default.

To do this the isoeditor factory API changes implement a WithEditor
function that is provided a callback function to run when an editor
is available to use.

WithEditor will block the calling thread until an editor is available
or the provided context times out.