MGMT-4693 Make minimal iso boot with UEFI (#1372)
* MGMT-4693 Compute the EFI load size correctly

Load Sectors * 2048 should be the size of efiboot.img rounded up
to a multiple of 2048.
This was preventing the uefi system from finding the boot image.

https://fedoraproject.org/wiki/User:Pjones/BootableCDsForBIOSAndUEFI#Testing_the_image

* MGMT-4693 Quote the rootfs url parameter in grub.cfg

Some special characters which are legal in a url are
used as separators in grub's syntax so we need quotes.

ref:
https://www.gnu.org/software/grub/manual/grub/html_node/Shell_002dlike-scripting.html

* MGMT-4693 Update the template version

This will force the service to regenerate the minimal
iso template with the new fixes.